Remove stale reverse pid mappings
Commit 3f72f54797cfc9d1f3bf8e00b4a1db0b9d246c57 seemed to accidentally
fix this problem partially. However, it did not go far enough to ensure
the reverse pid:group mapping was able to prevent unbounded growth.

If the API is used in full, i.e. joins are followed by leaves, rather
than relying on the pid going down then the mapping is maintained.

The behaviour found prior to 3f72f54797cfc9d1f3bf8e00b4a1db0b9d246c57
and this commit was that the overall size of the cpg state would grow
larger and larger over time with joins/leaves/'DOWN'/'EXIT' calls.

This meant that a) memory usage tended to increase and b) latency for
interacting with the pid:group mapping would also increase.